+++ src/hydrogen/module/session/keycloak.clj
@@ -3,21 +3,29 @@
             [duct.core.env :as env]
             [integrant.core :as ig]))
 
+(def ^:const options-defaults
+  {:keycloak {:realm (env/env '["KEYCLOAK_REALM" Str])
+              :url (env/env '["KEYCLOAK_URL" Str])
+              :client-id (env/env '["KEYCLOAK_CLIENT_ID" Str])}
+   :oidc {:issuer (env/env '["OIDC_ISSUER_URL" Str])
+          :audience (env/env '["OIDC_AUDIENCE" Str])
+          :jwks-uri (env/env '["OIDC_JWKS_URI" Str])}})
+
 (defn project-ns [config options]
   (:project-ns options (:duct.core/project-ns config)))
 
-(defn- session-config-base [project-ns]
+(defn- session-config-base [options project-ns]
   {(keyword (str project-ns ".api/config"))
    {:keycloak
-    {:realm (env/env '["KEYCLOAK_REALM" Str])
-     :url (env/env '["KEYCLOAK_URL" Str])
-     :clientId (env/env '["KEYCLOAK_CLIENT_ID" Str])}}
+    {:realm (get-in options [:keycloak :realm])
+     :url (get-in options [:keycloak :url])
+     :client-id (get-in options [:keycloak :client-id])}}
 
    :magnet.buddy-auth/jwt-oidc
    {:claims
-    {:iss (env/env '["OIDC_ISSUER_URL" Str])
-     :aud (env/env '["OIDC_AUDIENCE" Str])}
-    :jwks-uri (env/env '["OIDC_JWKS_URI" Str])}
+    {:iss (get-in options [:oidc :issuer])
+     :aud (get-in options [:oidc :audience])}
+    :jwks-uri (get-in options [:oidc :jwks-uri])}
 
    :duct.middleware.buddy/authentication
    {:backend :token
@@ -26,7 +34,7 @@
 
 (defn- session-config [options project-ns]
   (cond->
-   (session-config-base project-ns)
+   (session-config-base options project-ns)
 
     (:add-example-api? options)
     (assoc (keyword (str project-ns ".api/example"))
@@ -34,5 +42,6 @@
 
 (defmethod ig/init-key :hydrogen.module/session.keycloak [_ options]
   (fn [config]
-    (let [project-ns (project-ns config options)]
+    (let [project-ns (project-ns config options)
+          options (merge options-defaults options)]
       (core/merge-configs config (session-config options project-ns)))))